feat(cdk): create multiple stacks
`bin/cdk.ts` creates multiple stacks so that each stack corresponds to a
specific deployment stage (of "development" and "production"). If we
want to deploy a stack for a specific deployment stage, we can specify
the stack name corresponding to the deployment stage. It no longer uses
the context variable "dogs-business:deployment-stage".