chore(cdk): add npm scripts to facilitate cdk synth and deploy
The last commit changed the behavior of `cdk synth` and `cdk deploy`,
which used to synth and deploy the development stack, but now they synth
and deploy all the stacks; i.e., both for the development stage and the
production stage.

Introduces the following npm scripts so that I won't deploy the
production stack by mistake:
- `cdk-synth`: synthesizes the CloudFormation template for the
  development stage
- `cdk-synth:production`: synthesizes the CloudFormation template for
  the production stage
- `cdk-deploy`: deploy the CDK stack to the development stage
- `cdk-deploy:production`: deploy the CDK stack to the production stage

You have to configure the `BOOTSTRAP_QUALIFIER` environment variable
before runnint the above scripts.