feat(client): add function to check token expiration
Introduces a new function `isCognitoTokensExpiring` which checks if a
given `CognitoTokens` is expiring.

Defines the function in a new file `src/utils/passquito.ts`, because
`CognitoTokens` is part of the Passquito library. It might be better to
move the function to the Passquito library, but I have not decided yet.