feat(client): provide Resource API
The "authenticator-state" Pinia store injects the Dog's Business
Resource API provided from the upper layer. It uses the Resource API to
fetch user information.

In a new file `types/resource-api.ts`:
A new interface `ResourceApi` represents the Dog's Business Resource
API.

In a new file `providers/resource-api.ts`:
A new class `ResourceApiProvider` is a Vue plugin, which provides an
instance of `ResourceApi` to the lower layer.

In a new file `types/api-response.ts`:
A new interface `ApiResponse` is an extension to the `Response` of the
Fetch API, which can parse the body into the value of a specific type.

Updates the test cases for the "authenticator-state" Pinia store so that
they won't fail due to missing `ResourceApi`.