feat(gha): deploy CDK stack
Introduces a new GitHub Actions workflow `deploy-cdk` that is triggered
whenever a commit is pushed to the `main` branch and deploys the CDK
stack.

The workflow installs cargo-lambda and Zig so that cargo-lambda-cdk can
build Lambda runtimes without using Docker. It caches cargo artifacts.

`id-token: write` permission is required to assume the IAM role with
OIDC.

It requires the following secrets:
- `CDK_DEPLOY_ROLE_ARN`: ARN of the AWS IAM role that has a privilege to
  deploy the CDK stack to AWS
- `AWS_REGION`: AWS region to deploy the stack to
- `CDK_BOOTSTRAP_QUALIFIER`: bootstrap qualifier
- `PACKAGE_READER_PAT`: personal access token necessary to install
  packages from the GitHub package registry