feat(cdk): provision IAM role to read stack outputs
Introduces a new construct `StackReader` that provisions an IAM role
which can read outputs of the Dog's Business CloudFormation stack. This
IAM role is supposed to be assumed by a GitHub Actions workflow via
OIDC. `lib/oidc-config.ts` has to be customized for the actual GitHub
repository.

`CdkStack` defines an IAM principal to be assumed by GitHub Actions
workflow via OIDC. We can reuse the principal for other IAM roles.