feat(gha): add workflow to deploy client
Introduces a new GitHub Actions workflow `deploy-client` that builds and
uploads the client whenever a commit is pushed to the `main` branch. The
workflow is consisting of the following three jobs:
- `configure-client`
- `build-client`
- `deploy-client`

The `configure-client` job configures the .env file for the deployment
stage: development for now. The job is separated from the `build-client`
job because it needs to run in the `cdk` folder. It encrypts the .env
file and saves the encrypted file as a GitHub artifact so that the
`build-client` job can read it. The configuration script reads the
outputs from the Dog's Business CloudFormation stack.

The `build-client` job downloads and decrypts the .env file then builds
the contents in the `dist` folder. It archives and encrypts the `dist`
folder and saves the encrypted archive as a GitHub artifact so that the
`deploy-client` job can read it.

The `deploy-client` job downloads and decrypted the `dist` archive then
uploads it to the S3 bucket for the app distribution. The job obtains
the following parameters from the outputs of the Dog's Business
CloudFormation stack:
- name of the S3 bucket for the app distribution
- ARN of the IAM role that can upload contents to the S3 bucket

Artifacts are encrypted because they may contain sensitive information:
e.g., the URL of the development app distribution.

The workflow requires the following secrets:
- `STACK_OUTPUTS_READER_ROLE_ARN`: ARN of the IAM role that the workflow
  can assume to read outputs from the Dog's Business CloudFormation
  stack
- `ARTIFACT_ENCRYPTION_KEY`: secret key to encrypt artifacts. the key
  will be salted
- `PACKAGE_READER_PAT`: personal access token (PAT) to install packages
  from the GitHub npm registry
- `AWS_REGION`: AWS region to deploy