feat(gha): add configuration job
`deploy-client` workflow introduces a new job `configure-client` that
configures the `.env.{deployment-stage}` file for the client. Since the
configuration script is in the `cdk` folder, it has to be a separate
job with a specific Node setup.

It will encrypt the `.env.{deployment-stage}` file and store the
encrypted file as a GitHub artifact.