fix(gha): suppress stdout of cdk deploy
Suppresses the stdout of `cdk deploy` from being printed to the workflow
logs. Because the stack outputs may contain sensitive information: e.g.,
the URL of the development app distribution.

I was not able to figure out how to prevent specifically stack outputs
from being printed to the workflow logs:
- `--outputs-file` option did not work
- `--quite` option did not work